Switching back to ChaCha20Poly1305.
Just to put this to bed for myself, I'm pretty confident the current system has two key properties: key commitment and uniform distribution.

First, this design should be resistant to partitioning oracles by ensuring all AEAD keys are derived from KDF output seeded with, among other things, the public keys of all participants in the exchange. In "Message Franking via Committing Authentication Encryption", Grubbs et al write:

> If, however, one uses the input to the KDF as the opening, we can prove receiver binding assuming the KDF and MAC are collision resistant.

In all cases in this design, ChaCha20Poly1305 is seeded with the output of HKDF-SHA256, and the inputs to that include the sender's public key, the recipient's public key, and the ephemeral public key. Thus, all the identities are the opening, so receiver binding should hold.

Second, the Poly1305 tags should be uniformly distributed. The keys from which the Poly1305 keys are derived are HKDF-SHA256 output, which should be as uniformly distributed as can be managed. Therefore, the tags themselves should be indistinguishable from noise.